Секьюрити ошибка MyUserDetails при входе после авториации

+++ src/main/java/ru/work/graduatework/service/AdService.java
@@ -1,26 +1,34 @@
 package ru.work.graduatework.service;
 
 import com.sun.jdi.ObjectCollectedException;
+import lombok.RequiredArgsConstructor;
 import lombok.SneakyThrows;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.server.ResponseStatusException;
+import org.webjars.NotFoundException;
 import ru.work.graduatework.Entity.*;
 import ru.work.graduatework.dto.*;
 import ru.work.graduatework.mapper.AdMapper;
+import ru.work.graduatework.mapper.CommentMapper;
 import ru.work.graduatework.repository.AdRepository;
 import ru.work.graduatework.repository.CommentRepository;
 import ru.work.graduatework.repository.ImageRepository;
 import ru.work.graduatework.repository.UserRepository;
+import ru.work.graduatework.security.SecurityUtils;
 
+import java.time.Instant;
 import java.util.Collection;
 import java.util.List;
 
+import static ru.work.graduatework.security.SecurityUtils.checkPermissionToAds;
+import static ru.work.graduatework.security.SecurityUtils.getUserIdFromContext;
+
 @Service
+@RequiredArgsConstructor
 public class AdService {
 
     private final Logger logger = LoggerFactory.getLogger(AdService.class);
@@ -33,16 +41,8 @@ public class AdService {
     private final UserRepository userRepository;
     private final ImageService imageService;
     private final AdMapper adMapper;
+    private final CommentMapper commentMapper;
 
-    public AdService(AdRepository adRepository, CommentRepository commentRepository,
-                     UserRepository userRepository, ImageService imageService, ImageRepository imageRepository, AdMapper adMapper) {
-        this.adRepository = adRepository;
-        this.commentRepository = commentRepository;
-        this.userRepository = userRepository;
-        this.imageService = imageService;
-        this.imageRepository = imageRepository;
-        this.adMapper = adMapper;
-    }
 
     // Uses method - getAllAds    controller - AdsController
     public Collection<Ad> getAllAds() {
@@ -60,18 +60,18 @@ public class AdService {
     }
 
     @SneakyThrows
-    public Ad addAds(CreateAdDto createAdDto, MultipartFile adsImage, String Email) {
+    public Ad addAds(CreateAdDto createAdDto, MultipartFile adsImage) {
         logger.info("Current Method is - service AddAds");
-        Users user = userRepository.findByEmail(Email).orElseThrow();
-        Ad ad = adMapper.toEntity(createAdDto);
-        ad.setAuthor(user);
-        ad.setImage(imageService.uploadImage(adsImage));
-        return adRepository.save(ad);
+        Ad ads = adMapper.toEntity(createAdDto);
+        Users user = userService.getUserById(getUserIdFromContext());
+        ads.setAuthor(user);
+        ads.setImage(imageService.uploadImage(adsImage));
+        return adRepository.save(ads);
     }
 
 
-    public FullAdDto getFullAd(int id) {
-        return adMapper.toFullAdsDto(adRepository.findById(id).orElseThrow());
+    public FullAdDto getFullAd(long id) {
+        return adMapper.toFullAdsDto(adRepository.findById((int) id).orElseThrow());
     }
 
 
@@ -83,18 +83,18 @@ public class AdService {
     }
 
     // Uses method - updateAds    controller - AdsController
-    public Ad updateAds(int adId, CreateAdDto createAdDto) {
-        Ad ad = getAdsById(adId);
-        ad.setTitle(createAdDto.getTitle());
-        ad.setDescription(createAdDto.getDescription());
-        ad.setPrice(createAdDto.getPrice());
-        return adRepository.save(ad);
+    public Ad updateAds(long adId, CreateAdDto createAdDto) {
+        Ad ads = getAdsById(adId);
+        checkPermissionToAds(ads);
+        ads.setTitle(createAdDto.getTitle());
+        ads.setDescription(createAdDto.getDescription());
+        ads.setPrice(createAdDto.getPrice());
+        return adRepository.save(ads);
     }
 
     // Uses method - getAdsMe    controller - AdsController
-    public Collection<Ad> getAdsMe(String Email) {
-        Users user = userRepository.findByEmail(Email).orElseThrow();
-        return adRepository.findAllByAuthorId(user.getId());
+    public Collection<Ad> getAdsMe() {
+        return adRepository.findAllByAuthorId(getUserIdFromContext());
     }
 
     // Uses method - getComments    controller - AdsController
@@ -102,27 +102,38 @@ public class AdService {
         return commentRepository.findAllByAdId(adPk);
     }
 
+    public Comment addAdsComments(long adPk, AdCommentDto adsCommentDto) {
+        Comment adsComment = commentMapper.toEntity(adsCommentDto);
+        Users user = userService.getUserById(getUserIdFromContext());
+        adsComment.setAuthor(user);
+        adsComment.setAd(getAdsById(adPk));
+        adsComment.setCreatedAt(String.valueOf(Instant.now()));
+        return commentRepository.save(adsComment);
+    }
 
     public AdCommentDto getCommentsId(int ad_pk, int id) {
         Ad ad = adRepository.findById(ad_pk).orElseThrow();
         return new AdCommentDto(0, null, null);
     }
 
-    public void deleteCommentsId() {
+    public Comment deleteAdsComment(long adPk, long id) {
+        Comment comment = getAdsComment(adPk, id);
+        SecurityUtils.checkPermissionToAdsComment(comment);
+        commentRepository.delete(comment);
+        return comment;
     }
 
     // Uses method - updateComments    controller - AdsController
     public Comment updateComments(int adPk, int id, Comment commentUpdated) {
-
-        Comment comment = getAdsComment(adPk, id);
-        comment.setText(commentUpdated.getText());
-        return commentRepository.save(comment);
-
+        Comment adsComment = getAdsComment(adPk, id);
+        SecurityUtils.checkPermissionToAdsComment(adsComment);
+        adsComment.setText(commentUpdated.getText());
+        return commentRepository.save(adsComment);
     }
 
     // Uses method - updateAds
-    public Ad getAdsById(int asId) {
-        return adRepository.findById(asId).orElseThrow(
+    public Ad getAdsById(long asId) {
+        return adRepository.findById((int) asId).orElseThrow(
                 () -> new ResponseStatusException(
                         HttpStatus.NOT_FOUND, "The ad was not found"));
     }
@@ -140,37 +151,33 @@ public class AdService {
     // Uses method - getAdsComment    controller - AdsController
     public Comment getAdsComment(long adPk, long id) {
 
-        Comment comment = commentRepository.findByIdAndAdId(id, adPk)
-                .orElseThrow(() -> new ResponseStatusException(HttpStatus.NOT_FOUND,
-                        String.format("Ad %d " +
-                                "belonging to an ad with id %d not found", id, adPk)));
-        return comment;
+        return commentRepository.findByIdAndAdId(id, adPk)
+                .orElseThrow(() -> new NotFoundException(String.format("Comment with id %d " +
+                        "belonging to ad with id %d not found!", id, adPk)));
 
     }
 
-    // Uses method - deleteAdsComment    controller - AdsController
-    public Comment deleteAdsComment(long adPk, long id) {
-        Comment comment = getAdsComment(adPk, id);
-        commentRepository.delete(comment);
-        return comment;
-    }
 
     // Uses method - updateAdsImage    controller - AdsController
     @SneakyThrows
-    public void updateAdsImage(int id, MultipartFile image) {
-        Ad ad = getAdsById(id);
-
-        ad.setImage(imageService.uploadImage(image));
-        adRepository.save(ad);
+    public void updateAdsImage(long id, MultipartFile image) {
+        if (image == null) {
+            throw new NotFoundException("New ad image not uploaded");
+        }
+        Ad ads = getAdsById(id);
+        checkPermissionToAds(ads);
+        ads.setImage(imageService.uploadImage(image));
+        adRepository.save(ads);
     }
 
 
     // Uses method - removeAds    controller - AdsController
-    public Ad removeAdsById(int adId) {
-        Ad ad = getAdsById(adId);
+    public Ad removeAdsById(long adId) {
+        Ad ads = getAdsById(adId);
+        checkPermissionToAds(ads);
         commentRepository.deleteAdsCommentsByAdId(adId);
-        adRepository.delete(ad);
-        return ad;
+        adRepository.delete(ads);
+        return ads;
     }
